requirements.txt
hydra-core~=1.3.2
natsort~=8.4.0
numpy~=1.26.4
omegaconf~=2.3.0
open3d~=0.18.0
opencv_python~=4.9.0.80
Pillow~=10.3.0
torch~=2.3.0
torchvision~=0.18.0
rerun-sdk~=0.15.1